jarvis_min/requirements.txt
openai>=1.51.0,<2
python-dotenv>=1.0.1,<2
pydub>=0.25.1,<1
sounddevice>=0.4.6,<1
numpy>=2.3.3,<3
httpx==0.27.2
typer>=0.12.5,<1
colorama>=0.4.6,<1
loguru>=0.7.2,<1
